Update 103-fibonacci.c file

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	unsigned long num1 = 0, num2 = 2, next;
-	float sum;
+	unsigned long sum;
 
 	while (1)
 	{
